Refactor: Use system accent color for title bar caption buttons on Windows
This commit refactors the title bar caption buttons on Windows to use the system accent color when available.
If the system's accent color is enabled, the title bar buttons will use the accent color as their background.
Otherwise, they will use the default colors.

Also use `shellCloseColor` for Close button when the accent color is available.